+++ packages/framework/src/Component/Domain/Multidomain/MultidomainEntityClassFinderFacade.php
@@ -4,6 +4,7 @@ namespace Shopsys\FrameworkBundle\Component\Domain\Multidomain;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Shopsys\FrameworkBundle\Component\Doctrine\NotNullableColumnsFinder;
+use Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver;
 
 class MultidomainEntityClassFinderFacade
 {
@@ -27,16 +28,23 @@ class MultidomainEntityClassFinderFacade
      */
     protected $notNullableColumnsFinder;
 
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\EntityExtension\EntityNameResolver
+     */
+    protected $entityNameResolver;
+
     public function __construct(
         EntityManagerInterface $em,
         MultidomainEntityClassFinder $multidomainEntityClassFinder,
         MultidomainEntityClassProviderInterface $multidomainEntityClassProvider,
-        NotNullableColumnsFinder $notNullableColumnsFinder
+        NotNullableColumnsFinder $notNullableColumnsFinder,
+        EntityNameResolver $entityNameResolver
     ) {
         $this->em = $em;
         $this->multidomainEntityClassFinder = $multidomainEntityClassFinder;
         $this->multidomainEntityClassProvider = $multidomainEntityClassProvider;
         $this->notNullableColumnsFinder = $notNullableColumnsFinder;
+        $this->entityNameResolver = $entityNameResolver;
     }
 
     /**
@@ -58,7 +66,8 @@ class MultidomainEntityClassFinderFacade
     {
         $multidomainClassesMetadata = [];
         foreach ($this->getMultidomainEntitiesNames() as $multidomainEntityName) {
-            $multidomainClassesMetadata[] = $this->em->getMetadataFactory()->getMetadataFor($multidomainEntityName);
+            $resolvedClassName = $this->entityNameResolver->resolve($multidomainEntityName);
+            $multidomainClassesMetadata[] = $this->em->getMetadataFactory()->getMetadataFor($resolvedClassName);
         }
 
         return $this->notNullableColumnsFinder->getAllNotNullableColumnNamesIndexedByTableName($multidomainClassesMetadata);